Create WordBreak II.java
First run the WordBreak.java algorithm to get a boolean array to indicate which positions can be separated. Then use a recursive solution to check every possible separate position to get the final result. The boolean array will be calculated in O(n^2) time and the recursive solution will be sorted in O(2^m) where m is the number of the possible positions that the string s can be separated.