Create CloneGraph.java
The problem can be solved using BFS or DFS, the above solution is a DFS solution using the recursive call stack. The run time complexity if O(E) where E is the number of edges (each edge was visited only once). We need a map to store the mapping between nodes in the original graph and the new graph, the space complexity is O(V)