Create BinaryTreeMaximumSum.java
The key point of this problem is to use a class variable that track the total of all the subtree, while use recursion to calculate the total of the left and right subtree (return the max of left OR right subtree to upper level). Notice here that if both subtree has a negative sum, the value returned should be the root.val without adding neither left or right subtree.