Create DistinctSubsequence.java
The idea here is divide and conquer. using dynamic programming technique to avoid duplicate calculation and the running time is O(mn) where m and n is the length of the T and S. The matching can be decided this way. The number of distinct sequence between T.substring(0, i) and S.substring(0, j) will equal to the number of distinct sequence T.substring(0, i) and S.substring(0, j-1) if T.charAt(i) != S.charAt(j). Otherwise we need to add an additional T.substring(0, i-1), S.substring(0, j-1).