Create Symmetric Tree.java
If two trees are symmetric, then the root value of two tree must be the same, plus the left subtree of the first tree must be symmetric to the right subtree of the second tree, plus the right subtree of the first tree must be symmetric to the left subtree of the second tree. Two null trees are symmetric.