Create Longest Substring Without Repeating Characters.java
Use a hashtmap to keep track of the characters that was visited, use start and end pointer to track the longest substring without repeating chars. replace the start point when we find a duplicate char. Notice that if we have abcba, when we find the last a, we should keep start pointing to c, not moving it to b where the last a has appeared.